WIP Morph items.
Replaced all morph related data for player character actor in template.json with a single activeMorphId property.

Added a EPmorphModel class definition to define the schema for a morph item time, registered it as a dataModel for the "morph" type of Item.

Added a EPmorphSheet class definition for the Morph item sheet form, using a modified copy of templates/actor/partials/morph-details.html.

Added code to EPactor to support maintaining the activeMorphId property to always have a morph item attached, and to make it easily accessible within Handlebars or anywhere else.

Refactored a bunch of code in EPactor and EPactorSheet to support the new morph data structures for player characters.

Added a morph section in the template for the gear tab on the actor sheet, wired it to open the morph item sheet in a manner similar to the other items.

Allowed the actor sheet's Morph tab to be able to edit the active Morph item fields directly.

Allowed the actor sheet's Morph tab to be able to switch between morph items, including re-calculating all the actor's derived data.